Restructure KDA modules to Tokamax conventions
