+++ tokamax/_src/ops/experimental/kda/pallas_mosaic_tpu_fwd_kernel.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Pallas TPU forward kernels for experimental KDA."""
+"""Pallas/Mosaic TPU forward kernels for experimental KDA."""
 
 from __future__ import annotations
 
@@ -35,7 +35,7 @@ from tokamax._src.ops.experimental.kda.cp_utils import (
   _merge_initial_state,
   all_gather_into_tensor,
 )
-from tokamax._src.ops.experimental.kda.pallas_tpu_types import KdaResiduals
+from tokamax._src.ops.experimental.kda.pallas_mosaic_tpu_types import KdaResiduals
 from tokamax._src.ops.experimental.kda.utils import (
   _unalign_output,
   align_up,
